pyproject.toml
[tool.poetry]
name = "sqlnaut"
version = "0.1.0"
description = ""
authors = ["Joe Barrow <joseph.d.barrow@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fusepy = "^3.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
